+++ common/src/main/java/com/faboslav/structurify/common/checks/JigsawStructureBiomeCheck.java
@@ -4,7 +4,6 @@ import com.faboslav.structurify.common.config.data.StructureData;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderSet;
 import net.minecraft.core.QuartPos;
-import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.CheckerboardColumnBiomeSource;
 import net.minecraft.world.level.levelgen.Heightmap;
@@ -19,50 +18,52 @@ public final class JigsawStructureBiomeCheck
 		Structure.GenerationContext generationContext,
 		HolderSet<Biome> blacklistedBiomes
 	) {
-		var biomeCheckDistance = (int) Math.ceil(structureData.getBiomeCheckDistance() / 16.0);
+		int biomeCheckDistance = (int) Math.ceil(structureData.getBiomeCheckDistance() / 16.0);
 
 		if (biomeCheckDistance == 0 || generationContext.biomeSource() instanceof CheckerboardColumnBiomeSource) {
 			return true;
 		}
 
-		if(structureData.getBiomeCheckMode() == StructureData.BiomeCheckMode.BLACKLIST && structureData.getBiomeCheckBlacklistedBiomes().isEmpty()) {
+		var biomeCheckMode = structureData.getBiomeCheckMode();
+		if (biomeCheckMode == StructureData.BiomeCheckMode.BLACKLIST && structureData.getBiomeCheckBlacklistedBiomes().isEmpty()) {
 			return true;
 		}
 
-		ChunkPos chunkPos = generationContext.chunkPos();
+		var chunkPos = generationContext.chunkPos();
+		var biomeSource = generationContext.biomeSource();
+		var randomState = generationContext.randomState();
+		var biomeSampler = randomState.sampler();
+		int blockX = (chunkPos.x << 4);
+		int blockZ = (chunkPos.z << 4);
+
+		int surfaceY = generationContext.chunkGenerator().getFirstFreeHeight(
+			blockX,
+			blockZ,
+			Heightmap.Types.WORLD_SURFACE_WG,
+			generationContext.heightAccessor(),
+			generationContext.randomState()
+		);
 
 		for (int curChunkX = chunkPos.x - biomeCheckDistance; curChunkX <= chunkPos.x + biomeCheckDistance; curChunkX++) {
 			for (int curChunkZ = chunkPos.z - biomeCheckDistance; curChunkZ <= chunkPos.z + biomeCheckDistance; curChunkZ++) {
+				blockX = (curChunkX << 4) + 8;
+				blockZ = (curChunkZ << 4) + 8;
 
-				int blockX = curChunkX << 4;
-				int blockZ = curChunkZ << 4;
-
-				int surfaceY = generationContext.chunkGenerator().getFirstFreeHeight(
-					blockX,
-					blockZ,
-					Heightmap.Types.WORLD_SURFACE_WG,
-					generationContext.heightAccessor(),
-					generationContext.randomState()
-				);
-
-				Holder<Biome> biome = generationContext.biomeSource().getNoiseBiome(
+				Holder<Biome> biome = biomeSource.getNoiseBiome(
 					QuartPos.fromBlock(blockX),
-					QuartPos.fromBlock(surfaceY),
+					QuartPos.fromBlock(256),
 					QuartPos.fromBlock(blockZ),
-					generationContext.randomState().sampler()
+					biomeSampler
 				);
 
-				if (structureData.getBiomeCheckMode() == StructureData.BiomeCheckMode.STRICT && !generationContext.validBiome().test(biome)) {
-					return false;
+				switch (biomeCheckMode) {
+					case STRICT -> {
+						if (!generationContext.validBiome().test(biome)) return false;
+					}
+					case BLACKLIST -> {
+						if (blacklistedBiomes.contains(biome)) return false;
+					}
 				}
-
-				if (structureData.getBiomeCheckMode() == StructureData.BiomeCheckMode.BLACKLIST && blacklistedBiomes.contains(biome)) {
-					return false;
-				}
-
-				// TODO this check will not work for nether because of the Y level checks
-				//BlockPos blockPos = new BlockPos(blockX, surfaceY, blockZ);
-				//Structurify.getLogger().info("Structure biome " + biome + " at surface pos: " + blockPos);
 			}
 		}